End of refactoring
Завершён рефакторинг программы.
Описана логика работы.

Ожидает тестирования на ОС Linux
