+++ SSI.temp
@@ -1,0 +1,1 @@
+[2022/04/13 23:21:55] 3675523769448857384=31.3125
