+++ TempScanner.log
@@ -1,0 +1,14 @@
+[2022-04-13 23:21:48,950] DEBUG | RODOS_HID: Инициализация объекта класса RODOS_HID
+[2022-04-13 23:21:48,950] DEBUG | RODOS_HID: Поиск HID-устройства
+[2022-04-13 23:21:49,063] INFO | RODOS_HID: Найдено устройство [0x20a0:0x4173]
+[2022-04-13 23:21:49,091] INFO | RODOS_HID: Устройство "0x20a0:0x4173" успешно открыто
+[2022-04-13 23:21:49,093] DEBUG | __main__: Попытка найти конфигурационный файл
+[2022-04-13 23:21:49,094] DEBUG | __main__: Проверка наличия конфигурационного файла, переданных с аргументами из командной строки
+[2022-04-13 23:21:49,095] DEBUG | __main__: Поиск конфигурационного файла в стандартных местах расположения
+[2022-04-13 23:21:49,096] DEBUG | __main__: Проверка конфигурационного файла "D:\_GIT\Silines-temp\TempScanner_config.json"
+[2022-04-13 23:21:49,097] DEBUG | __main__: Конфигурационный файл "D:\_GIT\Silines-temp\TempScanner_config.json" не найден
+[2022-04-13 23:21:49,098] WARNING | __main__: Конфигурационные файлы не найдены. Будет создан новый: "D:\_GIT\Silines-temp\TempScanner_config.json"
+[2022-04-13 23:21:55,218] INFO | RODOS_HID: Найдено DALLAS - 1
+[2022-04-13 23:21:55,219] DEBUG | TemperatureScanner: Попытка сохранения конфигурационного файла.
+[2022-04-13 23:21:55,221] INFO | TemperatureScanner: Создан новый конфигурационный файл: "D:\_GIT\Silines-temp\TempScanner_config.json"
+[2022-04-13 23:21:55,530] INFO | TemperatureScanner: 3675523769448857384=31.3125
