+++ TempScanner_config.json
@@ -1,0 +1,9 @@
+{
+    "creation_date": "2022/04/13 23:21:55",
+    "last_edit_date": "2022/04/13 23:21:55",
+    "sensor_list": [
+        3675523769448857384
+    ],
+    "save_path": "D:\\_GIT\\Silines-temp\\SSI.temp",
+    "additional_log_path": []
+}
